# Optimize DistributedPubSub memory allocation