This change also results in a slight bump of performance. Note that, although this reduces _total_ memory allocation (Small Object Heap), it actually increases memory usage because it caches objects in memory.

This is the main 2 issues this PR trying to address:
![Original problem](https://github.com/user-attachments/assets/0f6c0116-5988-46a8-8d77-5b916abe30e9)

1. The recorded 1.1 Gb total SOH allocation by `ActorPath.Join()`
2. The recorded 575 Mb total SOH allocation by `ActorPath.op_Division()`

> [!NOTE]
> 
> Although the total bytes reported in this PR is quite scary (in the Gb and Mb range), note that this number represent worst case accumulated memory allocation of a stress benchmark application where DistributedPubSub is being subjected to a burst of more than 20,000,000 messages in a very short period of time.
>
> These are not memory leak, all of the allocated memories are reclaimed during GC.

## Changes

* Cache RegEx pattern string replace results
* Cache non-performant `ActorPath` `String.Join()` and immutable operations

## Checklist

For significant changes, please ensure that the following have been completed (delete if not relevant):

* [x] This change follows the [Akka.NET API Compatibility Guidelines](https://getakka.net/community/contributing/api-changes-compatibility.html).
* [x] This change follows the [Akka.NET Wire Compatibility Guidelines](https://getakka.net/community/contributing/wire-compatibility.html).

### Latest `dev` Benchmarks 

OSVersion:                         Microsoft Windows NT 6.2.9200.0
ProcessorCount:                    24
ClockSpeed:                        0 MHZ
Actor Count:                       48
Messages sent/received per client: 200000  (2e5)
Is Server GC:                      True
Thread count:                      55

| Num clients | Total [msg] | Msgs/sec | Total [ms] | Start Threads | End Threads |
|------------:|------------:|---------:|-----------:|--------------:|------------:|
|          1 |  200000 |    404041 |     495.04 |            55 |              67 |
|          5 | 1000000 |    401768 |    2489.68 |            72 |              88 |
|         10 | 2000000 |    413053 |    4842.76 |            92 |             119 |
|         15 | 3000000 |    425532 |    7050.14 |           123 |             118 |
|         20 | 4000000 |    433042 |    9237.34 |           123 |             115 |
|         25 | 5000000 |    417816 |   11967.61 |           119 |             114 |
|         30 | 6000000 |    429892 |   13957.77 |           119 |             115 |

### This PR's Benchmarks

OSVersion:                         Microsoft Windows NT 6.2.9200.0
ProcessorCount:                    24
ClockSpeed:                        0 MHZ
Actor Count:                       48
Messages sent/received per client: 200000  (2e5)
Is Server GC:                      True
Thread count:                      53

| Num clients | Total [msg] | Msgs/sec | Total [ms] | Start Threads | End Threads |
|------------:|------------:|---------:|-----------:|--------------:|------------:|
|          1 |  200000 |    505051 |     396.67 |            53 |              67 |
|          5 | 1000000 |    494560 |    2022.03 |            72 |              76 |
|         10 | 2000000 |    479847 |    4168.64 |            80 |              88 |
|         15 | 3000000 |    477403 |    6284.04 |            92 |             123 |
|         20 | 4000000 |    469374 |    8522.71 |           123 |             115 |
|         25 | 5000000 |    459348 |   10885.07 |           119 |             114 |
|         30 | 6000000 |    466636 |   12858.17 |           119 |             115 |