@Arkatufus 2025-05-14
Comparing the DPA recording of `dev` and this PR:

DEV:
![Original problem](https://github.com/user-attachments/assets/1b8c555e-04ff-4d54-ad76-eab07c6feb92)

PR:
![After Changes](https://github.com/user-attachments/assets/aebe6c83-0c71-4e5b-9724-7406c28d6daa)

We can see that `String.Join()` and `ActorPath.op_Division()` does not show up anymore.

Note that this is not a 100% fix, some of the old memory allocation has been shifted to `Utils.MakeKey()` instead (unavoidable), but comparing to the total 2 Gb total memory allocated by `String.Join()` and `ActorPath.op_Division()`, the `Utils.MakeKey()` 325 Mb allocation is a lot more acceptable.