@Arkatufus 2025-05-15
Here's the final benchmark for this PR:

OSVersion:                         Microsoft Windows NT 6.2.9200.0
ProcessorCount:                    24
ClockSpeed:                        0 MHZ
Actor Count:                       48
Messages sent/received per client: 200000  (2e5)
Is Server GC:                      True
Thread count:                      55

### dev

| Num clients | Total [msg] | Msgs/sec | Total [ms] | Start Threads | End Threads |
|------------:|------------:|---------:|-----------:|--------------:|------------:|
|          1 |  200000 |    628931 |     318.97 |            55 |              60 |
|          5 | 1000000 |    489237 |    2044.28 |            65 |              76 |
|         10 | 2000000 |    469264 |    4262.58 |            80 |              83 |
|         15 | 3000000 |    461326 |    6503.55 |            87 |             111 |
|         20 | 4000000 |    463554 |    8629.79 |           111 |             103 |
|         25 | 5000000 |    476690 |   10489.83 |           107 |             102 |
|         30 | 6000000 |    456032 |   13157.05 |           107 |             103 |

### Original PR

| Num clients | Total [msg] | Msgs/sec | Total [ms] | Start Threads | End Threads |
|------------:|------------:|---------:|-----------:|--------------:|------------:|
|          1 |  200000 |    649351 |     308.97 |            56 |              62 |
|          5 | 1000000 |    535332 |    1868.81 |            67 |              76 |
|         10 | 2000000 |    523287 |    3822.25 |            80 |              84 |
|         15 | 3000000 |    508906 |    5895.20 |            88 |             112 |
|         20 | 4000000 |    496155 |    8062.91 |           116 |             107 |
|         25 | 5000000 |    505051 |    9900.97 |           112 |             105 |
|         30 | 6000000 |    489796 |   12250.96 |           109 |             104 |

### readonly record struct

| Num clients | Total [msg] | Msgs/sec | Total [ms] | Start Threads | End Threads |
|------------:|------------:|---------:|-----------:|--------------:|------------:|
|          1 |  200000 |    689656 |     290.12 |            53 |              59 |
|          5 | 1000000 |    539666 |    1853.97 |            64 |              77 |
|         10 | 2000000 |    543331 |    3681.97 |            81 |              83 |
|         15 | 3000000 |    519481 |    5775.60 |            87 |             113 |
|         20 | 4000000 |    519751 |    7696.44 |           117 |             110 |
|         25 | 5000000 |    525763 |    9510.34 |           115 |             107 |
|         30 | 6000000 |    500793 |   11981.83 |           111 |             106 |

### non-concurrent non-static cache (final)

| Num clients | Total [msg] | Msgs/sec | Total [ms] | Start Threads | End Threads |
|------------:|------------:|---------:|-----------:|--------------:|------------:|
|          1 |  200000 |    598803 |     334.60 |            54 |              58 |
|          5 | 1000000 |    523561 |    1910.88 |            63 |              76 |
|         10 | 2000000 |    508647 |    3932.46 |            80 |              82 |
|         15 | 3000000 |    505391 |    5936.83 |            86 |             110 |
|         20 | 4000000 |    521309 |    7673.12 |           114 |             106 |
|         25 | 5000000 |    492320 |   10156.61 |           111 |             103 |
|         30 | 6000000 |    503694 |   11912.79 |           107 |             102 |