@Arkatufus 2025-05-15
Latest benchmark and DPA report comparison:

| Num clients | Total [msg] | Msgs/sec | Total [ms] | Start Threads | End Threads |
|------------:|------------:|---------:|-----------:|--------------:|------------:|
|          1 |  200000 |    687286 |     291.60 |            54 |              58 |
|          5 | 1000000 |    515199 |    1941.54 |            63 |              79 |
|         10 | 2000000 |    517197 |    3867.96 |            83 |              84 |
|         15 | 3000000 |    511859 |    5861.68 |            88 |             112 |
|         20 | 4000000 |    526109 |    7603.70 |           116 |             109 |
|         25 | 5000000 |    504185 |    9917.22 |           114 |             106 |
|         30 | 6000000 |    503356 |   11920.60 |           110 |             105 |

### dev:
![Original problem](https://github.com/user-attachments/assets/1b8c555e-04ff-4d54-ad76-eab07c6feb92)

### final PR:
![Final PR](https://github.com/user-attachments/assets/3b9e4dd6-5250-47b1-bc70-1778ee6d5c87)